# public/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Source Library API",
    "description": "Search and cite rare historical texts (alchemy, Hermeticism, Renaissance philosophy) from Source Library. All translations include DOI-backed academic citations.",
    "version": "1.1.0",
    "contact": {
      "name": "Source Library",
      "url": "https://sourcelibrary.org",
      "email": "team@sourcelibrary.org"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://sourcelibrary.org/api",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/dataset/v1/pages": {
      "get": {
        "summary": "Dataset: streaming page records (JSONL)",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Requires an API key; counts against the key daily page quota.",
        "parameters": [
          {
            "name": "language",
            "in": "query",
            "description": "Filter by language",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cluster",
            "in": "query",
            "description": "Taxonomy cluster",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "content",
            "in": "query",
            "description": "ocr | translation | both",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max 10000",
            "required": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Dataset API key (sl_data_...). Free at https://sourcelibrary.org/developers."
      }
    }
  }
}